move validation to a folder
