+++ services/frontend/src/i18n/tools/exportTransToCSV.js
@@ -20,10 +20,10 @@ const fs = require("fs");
 const path = require("path");
 const testHelpers = require("./validationHelperFunctions");
 
-const en = require("./en_CA.json");
-const fr = require("./fr_CA.json");
+const en = require("../en_CA.json");
+const fr = require("../fr_CA.json");
 
-const blacklistedKeys = require("./blacklistKeys.json");
+const blacklistedKeys = require("../blacklistKeys.json");
 
 /**
  * Format translations in JSON format for export
