+++ services/frontend/src/i18n/tools/validation.js
@@ -13,10 +13,10 @@
 
 /* eslint-disable no-console */
 const testHelpers = require("./validationHelperFunctions");
-const en = require("./en_CA.json");
-const fr = require("./fr_CA.json");
+const en = require("../en_CA.json");
+const fr = require("../fr_CA.json");
 
-const blacklistedKeys = require("./blacklistKeys.json");
+const blacklistedKeys = require("../blacklistKeys.json");
 
 /**
  * Main Code
